Move runtime error handling into interp/error.go; drop unused error from interp.New
interp.New never returned a non-nil error, so its signature simplifies to
just *Interpreter; Pool.grow and every call site follow. RuntimeError,
FrameInfo, escape, the sentinel error vars, and the panic/handler/unwind
helpers move out of interp.go into a dedicated error.go, mirroring the
types/error.go split already used elsewhere in the codebase.
