+++ benchmarks/jit_issue60_test.go
@@ -39,11 +39,11 @@ func BenchmarkJITIssue60(b *testing.B) {
 	for _, tt := range tests {
 		b.Run(tt.name, func(b *testing.B) {
 			b.Run("interp", func(b *testing.B) {
-				vm, _ := interp.New(tt.program, interp.WithThreshold(-1))
+				vm := interp.New(tt.program, interp.WithThreshold(-1))
 				runMiniVMProgram(b, vm, tt.want)
 			})
 			b.Run("jit", func(b *testing.B) {
-				vm, _ := interp.New(
+				vm := interp.New(
 					tt.program,
 					interp.WithTick(1),
 					interp.WithThreshold(1),
